feat: Add ContactDialog component and refactor component imports to absolute paths.
