+++ frontend/src/app/blog/page.tsx
@@ -3,8 +3,8 @@
 
 import { useState, useMemo } from "react";
 import Link from "next/link";
-import { Navbar } from "../../../components/layout/Navbar";
-import { Footer } from "../../../components/layout/Footer";
+import Navbar from "@/components/Navbar";
+import Footer from "@/components/Footer";
 import { getAllCategories, getAllPosts, getAllTags } from "@/lib/post";
 import { CategoryFilter } from "@/components/CategoryFilter";
 import { PostCard } from "@/components/ PostCard";
